requirements.txt
scipy>=1.2.1
numpy>=1.17.0
matplotlib>=3.1.2
opencv-python
torch>=1.2.0
torchvision>=0.4.0
tqdm>=4.60.0
Pillow>=8.2.0
h5py>=2.10.0
